+++ actionview/CHANGELOG.md
@@ -1,3 +1,6 @@
+*   Stop generating `<button>` elements with `[type="submit"]` and
+    `[name="button"]` by default.
 
+    *Sean Doyle*
 
 Please check [7-0-stable](https://github.com/rails/rails/blob/7-0-stable/actionview/CHANGELOG.md) for previous changes.